Update to support Lint output from SDK Tools r21.
Issues are now classified according to their Lint 'severity'
rather than their (relatively arbitrary) 'priority' value.

This means that upgrading the plugin will cause a slight change in
results for the first build executed, but it shouldn't be too major.

This also means that now you can mark a build as failed (or unstable)
if there are a certain number of 'Fatal' severity Lint issues found,
for example.

Finally, the dependency to the Lint JARs themselves was removed
as Lint now includes full issue information in its XML output.
This means that parsing a Lint file generated with r20 or older
will no longer show issue explanations directly within Jenkins.